+++ module-system/module-implementations/sov-blob-storage/tests/capability_tests.rs
@@ -164,10 +164,7 @@ fn do_deferred_blob_test(
             &SoftConfirmationsKernelGenesisConfig {
                 chain_state: ChainStateConfig {
                     current_time: Default::default(),
-                    gas_price_blocks_depth: 0,
-                    gas_price_maximum_elasticity: 0,
-                    initial_gas_price: GasPrice::ZEROED,
-                    minimum_gas_price: GasPrice::ZEROED,
+                    initial_base_fee_per_gas: GasPrice::ZEROED,
                 },
             },
             &mut kernel_working_set,
@@ -431,10 +428,7 @@ fn test_recovery_mode() {
             &SoftConfirmationsKernelGenesisConfig {
                 chain_state: ChainStateConfig {
                     current_time: Default::default(),
-                    gas_price_blocks_depth: 0,
-                    gas_price_maximum_elasticity: 0,
-                    initial_gas_price: GasPrice::ZEROED,
-                    minimum_gas_price: GasPrice::ZEROED,
+                    initial_base_fee_per_gas: GasPrice::ZEROED,
                 },
             },
             &mut kernel_working_set,
@@ -548,10 +542,7 @@ fn test_blobs_from_non_registered_sequencers_are_not_saved() {
             &BasicKernelGenesisConfig {
                 chain_state: ChainStateConfig {
                     current_time: Default::default(),
-                    gas_price_blocks_depth: 0,
-                    gas_price_maximum_elasticity: 0,
-                    initial_gas_price: GasPrice::ZEROED,
-                    minimum_gas_price: GasPrice::ZEROED,
+                    initial_base_fee_per_gas: GasPrice::ZEROED,
                 },
             },
             &mut kernel_working_set,
@@ -616,10 +607,7 @@ fn test_based_sequencing() {
             &BasicKernelGenesisConfig {
                 chain_state: ChainStateConfig {
                     current_time: Default::default(),
-                    gas_price_blocks_depth: 0,
-                    gas_price_maximum_elasticity: 0,
-                    initial_gas_price: GasPrice::ZEROED,
-                    minimum_gas_price: GasPrice::ZEROED,
+                    initial_base_fee_per_gas: GasPrice::ZEROED,
                 },
             },
             &mut kernel_working_set,