+++ module-system/module-implementations/sov-chain-state/src/tests.rs
@@ -8,10 +8,7 @@ fn test_config_serialization() {
     let time = Time::new(2, NanoSeconds::new(3).unwrap());
     let config = ChainStateConfig {
         current_time: time,
-        gas_price_blocks_depth: 10,
-        gas_price_maximum_elasticity: 5,
-        initial_gas_price: [2, 2].into(),
-        minimum_gas_price: [1, 1].into(),
+        initial_base_fee_per_gas: [2, 2].into(),
     };
 
     let data = r#"
@@ -20,10 +17,7 @@ fn test_config_serialization() {
             "secs":2,
             "nanos":3
         },
-        "gas_price_blocks_depth": 10,
-        "gas_price_maximum_elasticity": 5,
-        "initial_gas_price": [2, 2],
-        "minimum_gas_price": [1, 1]
+        "initial_base_fee_per_gas": [2, 2]
     }"#;
 
     let parsed_config: ChainStateConfig<TestSpec> = serde_json::from_str(data).unwrap();